requirements.txt
ray==2.4.0
click==8.1.3
biopython==1.81
grpcio==1.43.0
